# apps/idl/src/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Agent MVP API",
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    }
  ],
  "paths": {
...
    "/api/agent/runs": {
      "post": {
        "operationId": "createAgentRun",
        "summary": "Create an agent run",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentRunRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Agent run result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentRunResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApiErrorResponse": {
        "type": "object",
        "required": ["message"],
        "properties": {
          "message": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "details": {
            "type": "object",
            "additionalProperties": true
          }
        }
      },
...
      "AgentRunRequest": {
        "type": "object",
        "required": ["input"],
        "properties": {
          "conversationId": {
            "type": "string"
          },
          "input": {
            "type": "string",
            "minLength": 1
          },
          "pluginNames": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "metadata": {
            "type": "object",
            "additionalProperties": true
          },
          "pluginConfigs": {
            "type": "object",
            "additionalProperties": {
              "type": "object",
              "additionalProperties": true
            }
          }
        }
      },
      "AgentTraceEvent": {
        "type": "object",
        "required": ["id", "type", "message", "timestamp"],
        "properties": {
          "id": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": ["run.started", "plugin.started", "plugin.completed", "llm.delta", "run.completed", "run.failed"]
          },
          "message": {
            "type": "string"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "pluginName": {
            "type": "string"
          },
          "data": {
            "type": "object",
            "additionalProperties": true
          },
          "durationMs": {
            "type": "number"
          }
        }
      },
      "AgentRunResponse": {
        "type": "object",
        "required": ["runId", "conversationId", "output", "events"],
        "properties": {
          "runId": {
            "type": "string"
          },
          "conversationId": {
            "type": "string"
          },
          "output": {
            "type": "string"
          },
          "events": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AgentTraceEvent"
            }
          }
        }
      },
...
    }
  }
}