# packages/shared/src/generated/openapi.ts
// This file is generated by apps/idl/scripts/generate-openapi.ts.
// Do not edit it manually.

export type ApiErrorResponse = {
  "message": string;
  "code"?: string;
  "details"?: Record<string, unknown>;
};

export type HealthResponse = {
  "ok": boolean;
  "service": string;
  "version"?: string;
};

export type AgentPluginManifest = {
  "name": string;
  "description": string;
  "capabilities": string[];
  "enabled"?: boolean;
  "inputSchema"?: AgentPluginInputSchema;
  "defaultConfig"?: Record<string, unknown>;
};

export type AgentPluginInputSchema = {
  "minLength": number;
  "maxLength"?: number;
  "requiredMetadataKeys"?: string[];
};

export type AgentPluginListResponse = {
  "plugins": AgentPluginManifest[];
};

export type AgentRunRequest = {
  "conversationId"?: string;
  "input": string;
  "pluginNames"?: string[];
  "metadata"?: Record<string, unknown>;
  "pluginConfigs"?: Record<string, Record<string, unknown>>;
};

export type AgentTraceEvent = {
  "id": string;
  "type": "run.started" | "plugin.started" | "plugin.completed" | "llm.delta" | "run.completed" | "run.failed";
  "message": string;
  "timestamp": string;
  "pluginName"?: string;
  "data"?: Record<string, unknown>;
  "durationMs"?: number;
};

export type AgentRunResponse = {
  "runId": string;
  "conversationId": string;
  "output": string;
  "events": AgentTraceEvent[];
};

export type ConversationSummary = {
  "id": string;
  "title": string;
  "createdAt": string;
  "updatedAt": string;
};

export type ConversationMessage = {
  "id": string;
  "conversationId": string;
  "role": "user" | "assistant" | "system" | "tool";
  "content": string;
  "runId"?: string;
  "createdAt": string;
};

export type StoredAgentRun = {
  "id": string;
  "conversationId": string;
  "input": string;
  "output": string;
  "status": "running" | "completed" | "failed";
  "events": AgentTraceEvent[];
  "createdAt": string;
  "completedAt"?: string;
};

export type CreateConversationRequest = {
  "title"?: string;
};

export type ConversationListResponse = {
  "conversations": ConversationSummary[];
};

export type ConversationDetailResponse = {
  "conversation": ConversationSummary;
  "messages": ConversationMessage[];
  "runs": StoredAgentRun[];
};

export type ConversationMessagesResponse = {
  "messages": ConversationMessage[];
};

export type DeleteConversationResponse = {
  "ok": boolean;
};

export type GetHealthOperationRequest = undefined;
export type GetHealthOperationResponse = HealthResponse;

export type ListAgentPluginsOperationRequest = undefined;
export type ListAgentPluginsOperationResponse = AgentPluginListResponse;

export type CreateAgentRunOperationRequest = AgentRunRequest;
export type CreateAgentRunOperationResponse = AgentRunResponse;

export type ListConversationsOperationRequest = undefined;
export type ListConversationsOperationResponse = ConversationListResponse;

export type CreateConversationOperationRequest = CreateConversationRequest;
export type CreateConversationOperationResponse = ConversationDetailResponse;

export type GetConversationOperationRequest = undefined;
export type GetConversationOperationResponse = ConversationDetailResponse;

export type DeleteConversationOperationRequest = undefined;
export type DeleteConversationOperationResponse = DeleteConversationResponse;

export type ListConversationMessagesOperationRequest = undefined;
export type ListConversationMessagesOperationResponse = ConversationMessagesResponse;